move StreamConsumer to libs-scala/grpc-test-utils
