+++ ledger-test-tool/infrastructure/BUILD.bazel
@@ -28,9 +28,9 @@ load("//ledger-test-tool:conformance.bzl", "testtool_lf_versions")
             "//ledger/ledger-api-errors",
             "//libs-scala/test-evidence/tag:test-evidence-tag",
             "//test-common:dar-files-%s-lib" % lf_version,
-            "//test-common:test-common-%s" % lf_version,
             "//libs-scala/build-info",
             "//libs-scala/contextualized-logging",
+            "//libs-scala/grpc-test-utils",
             "//libs-scala/resources",
             "//libs-scala/resources-grpc",
             "//libs-scala/resources-pekko",
