+++ libs-scala/grpc-test-utils/src/main/scala/com/daml/grpc/test/ObserverFilter.scala
@@ -1,11 +1,11 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.testing
+package com.daml.grpc.test
 
 import io.grpc.stub.StreamObserver
 
-private[testing] final class ObserverFilter[A](predicate: A => Boolean)(delegate: StreamObserver[A])
+private[test] final class ObserverFilter[A](predicate: A => Boolean)(delegate: StreamObserver[A])
     extends StreamObserver[A] {
 
   override def onNext(value: A): Unit = synchronized {
