+++ libs-scala/grpc-test-utils/src/main/scala/com/daml/grpc/test/TimeBoundObserver.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.testing
+package com.daml.grpc.test
 
 import com.daml.timer.Delayed
 import io.grpc.Context
