+++ observability/metrics/BUILD.bazel
@@ -85,6 +85,7 @@ da_scala_test_suite(
         ":metrics",
         ":metrics-test-lib",
         "//libs-scala/concurrent",
+        "//libs-scala/grpc-test-utils",
         "//libs-scala/ledger-resources",
         "//libs-scala/ledger-resources:ledger-resources-test-lib",
         "//libs-scala/resources",
@@ -93,7 +94,6 @@ da_scala_test_suite(
         "//libs-scala/rs-grpc-pekko:rs-grpc-pekko-tests-lib",
         "//libs-scala/sample-service",
         "//libs-scala/testing-utils",
-        "//test-common",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:io_dropwizard_metrics_metrics_core",
         "@maven//:io_grpc_grpc_api",
